requirements.txt
apache-airflow
httpx
requests
pydantic
pydantic-settings
python-dateutil
tenacity
structlog
asyncio-pool
pytest
pytest-asyncio
pytest-cov
pytest-mock
black
isort
flake8
mypy
python-dotenv
cryptography
SQLAlchemy
PyMySQL